backend/auth_service/requirements.txt
Flask
python-dotenv
boto3
PyJWT
requests-oauthlib
